Refactor: Rewrote the application.
- TODO: Replace the Service.Storage with an interface called Storer
