+++ pkg/acache/routes.go
@@ -3,8 +3,6 @@ package acache
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 )
 
 // Route Route
@@ -15,20 +13,6 @@ type Route struct {
 	Data        []byte `json:"data"`
 	ContentType string `json:"contentType"`
 }
-
-//ContainsURL ContainsURL returns true if the slice of routes contains an URL
-func (routes *Routes) ContainsURL(url string) (bool, error) {
-	for _, v := range *routes {
-		if v.URL == url {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-// Routes Routes
-type Routes []Route
 
 // NewRouteFromBytes RouteFromBytes...
 func NewRouteFromBytes(bytes []byte) (Route, error) {
@@ -41,6 +25,19 @@ func NewRouteFromBytes(bytes []byte) (Route, error) {
 	return routes, nil
 }
 
+// Routes Routes
+type Routes []Route
+
+//ContainsURL ContainsURL returns true if the slice of routes contains an URL
+func (routes *Routes) ContainsURL(url string) (bool, error) {
+	for _, v := range *routes {
+		if v.URL == url {
+			return true, nil
+		}
+	}
+
+	return false, nil
+}
 func (routes Routes) ToString() string {
 	var output string
 	for i, v := range routes {
@@ -59,14 +56,4 @@ func (routes *Routes) PrintInfo() {
 	for i, v := range *routes {
 		fmt.Printf("%d) %s\n\tAlias: %s\n\tKey: %s\n\tContent-Type: %s\n", i, v.URL, v.Alias, v.ID, v.ContentType)
 	}
-}
-
-func fetchItem(url string) ([]byte, *http.Response, error) {
-	res, err := http.Get(url)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	body, err := ioutil.ReadAll(res.Body)
-	return body, res, err
 }
